make the upstream bazel sandbox work with the custom rootfs
Summary:
Sandbox 2.0 uses the following model:
- create a unique directory to be the sandbox root
- create a working directory inside the sandbox root
- create symlinks for input files in the working directory
- mount / at the sandbox root - make this read-only
- mount the working directory onto itself - make this read-write
- copy outputs from the working directory

This results in far fewer mounts, but there was no way I could get it to work with a custom rootfs. The problem is due to local repository rules, the input files can come from anywhere (`/usr/local/encap/zookeeper` is one such location), and the only way to ensure that those directories are correctly mounted into the custom rootfs is to mount the individual input files, which is what sandbox 1.0 did. So, we now do the following:
- create a unique directory to be the sandbox root
- create a working directory inside the sandbox root
- mount system directories inside the sandbox root - make these read-only
- mount inputs inside the sandbox root - make these read-only
- mount the working directory onto itself - make this read-write
- copy outputs from the working directory

Test Plan:
was able to test `/srv/bazel/output/bazel test --spawn_strategy=sandboxed --genrule_strategy=sandboxed` the following:
- `//dropbox/devtools/dbshell/tests:dbshell_test`
- `//go/src/dropbox/storage/blockstore:blockstore_test`

Reviewers: anupc, benjamin

Subscribers: wwu, anupc

Differential Revision: https://tails.corp.dropbox.com/D265719