cherry-pick: sandbox should create regular empty files, not symlink to /dev/null.
Summary:
upstream commit: 738e5727bab23961c7cc423606f38f8b97a8150e
motivation: Now that we are trying to switch to a model where everything in the sandbox exec root is a symlink, some binaries need to be able to make a full copy of their runfiles folder (like, anything using java). This doesn't work well when there are symlinks to /dev/null

original message:
Adds a test based on our Python rules that makes sure that this actually
fixes the issue. Thanks to @duggelz for the suggestion. It seems like
our Python rules are the only place that actually provides an
EmptyFilesSupplier to Runfiles, so there's probably no simpler way to
test this behavior in an integration test.

--
PiperOrigin-RevId: 148656193
MOS_MIGRATED_REVID=148656193

Reviewers: anupc, benjamin

Reviewed By: anupc

Differential Revision: https://tails.corp.dropbox.com/D282649